requirements.txt
# Core Streamlit and Web Framework
streamlit==1.45.1
streamlit-folium==0.25.3
altair==5.5.0
contextily==1.7.0

# Data Processing and Analysis
pandas==2.2.3
numpy==2.2.6
geopandas==0.14.0

# Google Cloud Storage
google-cloud-storage==3.1.0
google-cloud-core==2.4.2
google-auth==2.38.0
google-api-core==2.24.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1

# Data Validation and Serialization
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4

# Visualization
plotly
pydeck==0.9.1
Pillow==11.2.1

# HTTP and Networking
requests==2.32.3
httpx==0.27.2
httpcore==1.0.6
aiohttp==3.12.0
httpx-sse==0.4.0

# Utilities
python-dateutil
python-dotenv==1.1.0
pytz==2025.2
certifi==2024.8.30
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0

# Git Integration
GitPython==3.1.44
gitdb==4.0.12

# JSON and Data Structures
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
orjson==3.10.18
dataclasses-json==0.6.7


# Markdown and Templating
Jinja2==3.1.6
MarkupSafe==3.0.2

# Configuration and Settings
PyYAML==6.0.2
toml==0.10.2

# Async and Concurrency
aiosignal==1.3.2
async-timeout==4.0.3
anyio==4.6.2.post1

# Caching and Storage
cachetools==5.5.2

# Progress and Logging
tqdm==4.66.5

# Other Dependencies
tenacity==9.1.2
watchdog==6.0.0
packaging==24.2
regex==2024.11.6
SQLAlchemy==2.0.41

# Testing (optional, can be removed for production)
pytest
pytest-mock